+++ src/Relationship/OneToOne.php
@@ -46,7 +46,7 @@ class OneToOne extends DeletableRelationship
         Record $nativeRecord,
         array &$foreignRecords
     ) : void {
-        $nativeRecord->{$this->name} = false;
+        $nativeRecord->{$this->name} = NotLoaded::getInstance();
         foreach ($foreignRecords as $index => $foreignRecord) {
             if ($this->recordsMatch($nativeRecord, $foreignRecord)) {
                 $nativeRecord->{$this->name} = $foreignRecord;